@@ -11,7 +11,7 @@ buildscript {
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.6.3'
     }
 }
 apply plugin: 'com.android.application'
@@ -40,13 +40,14 @@ repositories {
 
 dependencies {
     implementation  group: 'cz.msebera.android' , name: 'httpclient' , version: '4.4.+'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:support-v13:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.test.ext:junit:1.1.1'
     implementation 'com.google.gms:google-services:3.2.1'
     implementation 'com.google.android.gms:play-services-gcm:11.0.4'
     implementation 'com.google.code.findbugs:jsr305:3.0.2'
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.jakewharton:butterknife:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.0.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.0.0'
     implementation 'com.karumi:dexter:5.0.0'
     implementation 'ch.acra:acra:4.5.0'
     implementation 'commons-net:commons-net:3.6'
@@ -54,7 +55,7 @@ dependencies {
     implementation 'com.infstory:proguard-annotations:1.0.1'
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
     implementation 'org.osmdroid:osmdroid-android:6.1.6'
-    implementation "com.facebook.react:react-native:0.52.2" // From node_modules
+    implementation "com.facebook.react:react-native:0.62.2" // From node_modules
 }
 
 android {
@@ -67,12 +68,12 @@ android {
         abortOnError false
     }
 
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "gpsplus.rtkgps"
-        minSdkVersion 24
-        targetSdkVersion 27
+        minSdkVersion 27
+        targetSdkVersion 28
         multiDexEnabled true
         ndk {
             abiFilters "armeabi-v7a", "x86"
